Refactor dataset/signal endpoints into separate module
