+++ lilac/server_dataset_test.py
@@ -20,8 +20,6 @@ from .data.dataset_test_utils import (
 from .router_dataset import (
   AddLabelsOptions,
   Column,
-  ComputeSignalOptions,
-  DeleteSignalOptions,
   SelectRowsOptions,
   SelectRowsResponse,
   SelectRowsSchemaOptions,
@@ -286,30 +284,6 @@ def test_select_rows_schema_no_cols() -> None:
       }
     )
   )
-
-
-def test_compute_signal_auth(mocker: MockerFixture) -> None:
-  mocker.patch.dict(os.environ, {'LILAC_AUTH_ENABLED': 'True'})
-
-  url = f'/api/v1/datasets/{TEST_NAMESPACE}/{TEST_DATASET_NAME}/compute_signal'
-  response = client.post(
-    url, json=ComputeSignalOptions(signal=LengthSignal(), leaf_path=('people', 'name')).model_dump()
-  )
-  assert response.status_code == 401
-  assert response.is_error is True
-  assert 'User does not have access to compute signals over this dataset.' in response.text
-
-
-def test_delete_signal_auth(mocker: MockerFixture) -> None:
-  mocker.patch.dict(os.environ, {'LILAC_AUTH_ENABLED': 'True'})
-
-  url = f'/api/v1/datasets/{TEST_NAMESPACE}/{TEST_DATASET_NAME}/delete_signal'
-  response = client.request(
-    'DELETE', url, json=DeleteSignalOptions(signal_path=('doesnt', 'matter')).model_dump()
-  )
-  assert response.status_code == 401
-  assert response.is_error is True
-  assert 'User does not have access to delete this signal.' in response.text
 
 
 def test_update_settings_auth(mocker: MockerFixture) -> None:
